+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Added:
 
 - Added `Open URL` entry to URL context menus
 - Setting to not hide consecutive if previous message was a image preview
+- Channel discovery buffer to view all channels on a server (using LIST). Can be opened via sidebar button or command bar.
 - Setting to open a new pane, replace current pane, or open a new window after `/join`
 - Emoji, channel, and user completion can be used with commands
 - Auto-completion (emoji/channel/user) will auto-complete at the text cursor (instead of the end of the current input)